re-move tax-placholder.js to lib/tax
