+++ client/lib/tax/index.js
@@ -1,7 +1,7 @@
 /** @format */
 
 import { parse as parseUrl } from 'url';
-import { get, includes, isString, mapValues, merge, pickBy } from 'lodash';
+import { get, includes, isString, mapValues, merge, pickBy, map, has } from 'lodash';
 /**
  * Internal dependencies
  */
@@ -155,3 +155,22 @@ const automap = f => arg => ( arg && typeof arg.map === 'function' ? arg.map( f 
 export const injectPurchasesWithPlaceholderValues = automap(
 	injectSinglePurchaseWithPlaceholderValues
 );
+
+export function maybeInjectPlaceholderTaxAmountIntoCharge( charge ) {
+	if (
+		config.isEnabled( 'show-tax' ) &&
+		! has( charge, 'tax_amount' ) &&
+		isString( get( charge, 'amount' ) )
+	) {
+		return Object.assign( charge, {
+			tax_amount: charge.amount.replace( /(\d+)(\d)\.(\d*)\d/, '$1.$2$3' ),
+		} );
+	}
+	return charge;
+}
+
+export function maybeInjectPlaceholderTaxAmountIntoCharges( charges ) {
+	return config.isEnabled( 'show-tax' )
+		? map( charges, maybeInjectPlaceholderTaxAmountIntoCharge )
+		: charges;
+}
